Time: 0 ms (100.00%) | Memory: 71.8 MB (51.61%) - LeetSync

+++ 53-maximum-subarray/maximum-subarray.cpp
@@ -1,9 +1,9 @@
 class Solution {
 public:
     int maxSubArray(vector<int>& nums) {
-        long long n = nums.size();
-        long long sum = 0;
-        long long maxi = INT_MIN;
+        int n = nums.size();
+        int sum = 0;
+        int maxi = INT_MIN;
         for(int i=0;i<n;i++)
         {
             sum = sum + nums[i];
